+++ numpydoc/validate.py
@@ -5,13 +5,19 @@ Analyze docstrings to detect errors.
 Call ``validate(object_name_to_validate)`` to get a dictionary
 with all the detected errors.
 """
+
+from copy import deepcopy
+from typing import Set
 import ast
 import collections
 import importlib
 import inspect
+import os
 import pydoc
 import re
 import textwrap
+import tokenize
+
 from .docscrape import get_doc_object
 
 
@@ -100,6 +106,72 @@ ERROR_MSGS = {
 
 # Ignore these when evaluating end-of-line-"." checks
 IGNORE_STARTS = (" ", "* ", "- ")
+
+# inline comments that can suppress individual checks per line
+IGNORE_COMMENT_PATTERN = re.compile("(?:.* numpydoc ignore[=|:] ?)(.+)")
+
+
+def extract_ignore_validation_comments(filepath: os.PathLike) -> dict[int, list[str]]:
+    """
+    Extract inline comments indicating certain validation checks should be ignored.
+
+    Parameters
+    ----------
+    filepath : os.PathLike
+        Path to the file being inspected.
+
+    Returns
+    -------
+    dict[int, list[str]]
+        Mapping of line number to a list of checks to ignore.
+    """
+    with open(filepath) as file:
+        numpydoc_ignore_comments = {}
+        last_declaration = 1
+        declarations = ["def", "class"]
+        for token in tokenize.generate_tokens(file.readline):
+            if token.type == tokenize.NAME and token.string in declarations:
+                last_declaration = token.start[0]
+            if token.type == tokenize.COMMENT:
+                match = re.match(IGNORE_COMMENT_PATTERN, token.string)
+                if match:
+                    rules = match.group(1).split(",")
+                    numpydoc_ignore_comments[last_declaration] = rules
+    return numpydoc_ignore_comments
+
+
+def get_validation_checks(validation_checks: Set[str]) -> Set[str]:
+    """
+    Get the set of validation checks to report on.
+
+    Parameters
+    ----------
+    validation_checks : set[str]
+        A set of validation checks to report on. If the set is ``{"all"}``,
+        all checks will be reported. If the set contains just specific checks,
+        only those will be reported on. If the set contains both ``"all"`` and
+        specific checks, all checks except those included in the set will be
+        reported on.
+
+    Returns
+    -------
+    set[str]
+        The set of validation checks to report on.
+    """
+    valid_error_codes = set(ERROR_MSGS.keys())
+    if "all" in validation_checks:
+        block = deepcopy(validation_checks)
+        validation_checks = valid_error_codes - block
+
+    # Ensure that the validation check set contains only valid error codes
+    invalid_error_codes = validation_checks - valid_error_codes
+    if invalid_error_codes:
+        raise ValueError(
+            f"Unrecognized validation code(s) in numpydoc_validation_checks "
+            f"config value: {invalid_error_codes}"
+        )
+
+    return validation_checks
 
 
 def error(code, **kwargs):
@@ -506,9 +578,15 @@ def validate(obj_name, validator_cls=None, **validator_kwargs):
     else:
         doc = validator_cls(obj_name=obj_name, **validator_kwargs)
 
+    # module docstring will be lineno 0, which we change to 1 for readability of the output
+    ignore_validation_comments = extract_ignore_validation_comments(
+        doc.source_file_name
+    ).get(doc.source_file_def_line or 1, [])
+
     errs = []
     if not doc.raw_doc:
-        errs.append(error("GL08"))
+        if "GL08" not in ignore_validation_comments:
+            errs.append(error("GL08"))
         return {
             "type": doc.type,
             "docstring": doc.clean_doc,
@@ -630,6 +708,9 @@ def validate(obj_name, validator_cls=None, **validator_kwargs):
 
     if not doc.examples:
         errs.append(error("EX01"))
+
+    errs = [err for err in errs if err[0] not in ignore_validation_comments]
+
     return {
         "type": doc.type,
         "docstring": doc.clean_doc,
